replace debug boolean with dropdown
Instead of halting every step in CI when debugging via tmate is enabled,
allow the actor to select which step they want to debug.

Previously, if a dev was trying to debug the test step, for example,
they would have to first tmate into the build step and close the session
to allow the build to proceed.

Ref #74